+++ models/tt_transformers/demo/voxtral_4b_tts_2603/_stubs/feed_forward.py
@@ -192,8 +192,8 @@ def build(device, torch_module):
         # HiFi2 on the bf8_b weights (two phases cover a bf8_b mantissa); the hidden activation is
         # handed over in L1 and in bf16, since the down projection multicasts it whole.
         gated = ttnn.multiply(
-            _lin(h, w1, dtype=ttnn.float32, compute_kernel_config=_TALL_COMPUTE, memory_config=ttnn.L1_MEMORY_CONFIG),
-            _lin(h, w3, dtype=ttnn.float32, compute_kernel_config=_TALL_COMPUTE, memory_config=ttnn.L1_MEMORY_CONFIG),
+            _lin(h, w1, dtype=ttnn.bfloat16, compute_kernel_config=_TALL_COMPUTE, memory_config=ttnn.L1_MEMORY_CONFIG),
+            _lin(h, w3, dtype=ttnn.bfloat16, compute_kernel_config=_TALL_COMPUTE, memory_config=ttnn.L1_MEMORY_CONFIG),
             input_tensor_a_activations=[ttnn.UnaryOpType.SILU],
             dtype=ttnn.bfloat16,
             memory_config=ttnn.L1_MEMORY_CONFIG,
